refactor : reduce redundant code
